toggleFavorite function via state uplift in _app.js
