+++ components/Card/Card.js
@@ -1,15 +1,12 @@
 import styled from "styled-components";
 import { IoStar, IoLocationSharp } from "react-icons/io5";
-import FavoriteButton from "../FavoriteButton/FavoriteButton";
+
 import Link from "next/link";
 
-export default function Card({ location, onToggleFavorite }) {
+export default function Card({ location }) {
   return (
     <StyledLink href={`/${location.slug}`}>
       <StyledLi image={location.image}>
-        <StyledButton onClick={() => onToggleFavorite(location.id)}>
-          <FavoriteButton />
-        </StyledButton>
         <StyledTitle>{location.title}</StyledTitle>
         <StyledRating>
           <IoStar style={{ color: `var(--color-star)` }} />
@@ -49,20 +46,11 @@ const StyledLi = styled.li`
   grid-template-columns: repeat(4, 1fr);
   grid-template-rows: auto;
   grid-template-areas:
-    "favoriteButton title title title"
+    "title title title title"
     ". . rating rating "
     ". tags tags tags"
     ". city city city";
   align-items: center;
-`;
-
-const StyledButton = styled.button`
-  background: none;
-  border: none;
-  grid-area: favoriteButton;
-  color: var(--color-foreground-alt);
-  justify-self: flex-start;
-  z-index: 1;
 `;
 
 const StyledTitle = styled.h2`
