--- pages/_app.tsx
@@ -1,11 +1,0 @@
-import type { AppProps } from "next/app";
-import GlobalStyles from "../components/GlobalStyles";
-
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <>
-      <GlobalStyles />
-      <Component {...pageProps} />
-    </>
-  );
-}
